+++ schedule-priority/schedule-priority.py
@@ -1,5 +1,0 @@
-# Anki Schedule prioritization
-# Author Ricardo Saturnino
-# Version: 2.0
-
-